@@ -1,4 +1,5 @@
 ﻿#region LICENSE
+
 // PaperMalKing.
 // Copyright (C) 2021 N0D4N
 // 
@@ -14,13 +15,14 @@
 // 
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 #endregion
 
 namespace PaperMalKing.AniList.Wrapper.GraphQL
 {
-    internal static class Queries
-    {
-        public const string GetUserInitialInfoByUsernameQuery = @"
+	internal static class Queries
+	{
+		public const string GetUserInitialInfoByUsernameQuery = @"
           query ($username: String, $favouritePage: Int) {
             User(name: $username) {
               id
@@ -70,7 +72,7 @@ namespace PaperMalKing.AniList.Wrapper.GraphQL
           }
           ";
 
-        public const string FavouritesInfoQuery = @"
+		public const string FavouritesInfoQuery = @"
           query ($page: Int, $animeIds: [Int], $mangaIds: [Int], $charIds: [Int], $staffIds: [Int], $studioIds: [Int]) {
             Animes: Page(page: $page, perPage: 50) {
               pageInfo{
@@ -193,5 +195,5 @@ namespace PaperMalKing.AniList.Wrapper.GraphQL
             }
           }
           ";
-    }
+	}
 }
