@@ -17,16 +17,17 @@
 #endregion
 
 using System;
-using System.Diagnostics.CodeAnalysis;
+using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
 namespace PaperMalKing.AniList.Wrapper.Models
 {
-	[SuppressMessage("Naming", "CA1711")]
+	#pragma warning disable CA1711
 	public sealed class MediaListCollection
+	#pragma warning restore CA1711
     {
         [JsonPropertyName("lists")]
-        public MediaListGroup[] Lists { get; init; } = Array.Empty<MediaListGroup>();
+        public IReadOnlyList<MediaListGroup> Lists { get; init; } = Array.Empty<MediaListGroup>();
 
         public static readonly MediaListCollection Empty = new ();
 	}
