@@ -18,7 +18,7 @@
 
 using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
+using System.ComponentModel;
 using System.Globalization;
 using System.Linq;
 using System.Text.Json.Serialization;
@@ -28,7 +28,7 @@ namespace PaperMalKing.AniList.Wrapper.Models
 {
     public sealed class MediaListEntry
     {
-        private readonly Dictionary<ScoreFormat, byte> _scores = new();
+        private readonly Dictionary<ScoreFormat, byte> _scores = new(4);
 
         [JsonPropertyName("status")]
         public MediaListStatus Status { get; init; }
@@ -39,35 +39,34 @@ namespace PaperMalKing.AniList.Wrapper.Models
         [JsonPropertyName("notes")]
         public string? Notes { get; init; }
 
-		[SuppressMessage("", "CA2227")]
 		[JsonPropertyName("advancedScores")]
         public Dictionary<string, float>? AdvancedScores { get; init; }
 
         [JsonPropertyName("point100Score")]
         public byte Point100Score
         {
-            [Obsolete("", true)] get => throw new NotSupportedException();
+            [Obsolete("", true), EditorBrowsable(EditorBrowsableState.Never)] get => throw new NotSupportedException();
             init => this._scores.Add(ScoreFormat.POINT_100, value);
         }
 
         [JsonPropertyName("point10Score")]
         public byte Point10Score
         {
-            [Obsolete("", true)] get => throw new NotSupportedException();
+            [Obsolete("", true), EditorBrowsable(EditorBrowsableState.Never)] get => throw new NotSupportedException();
             init => this._scores.Add(ScoreFormat.POINT_10, value);
         }
 
         [JsonPropertyName("point5Score")]
         public byte Point5Score
         {
-            [Obsolete("", true)] get => throw new NotSupportedException();
+            [Obsolete("", true), EditorBrowsable(EditorBrowsableState.Never)] get => throw new NotSupportedException();
             init => this._scores.Add(ScoreFormat.POINT_5, value);
         }
 
         [JsonPropertyName("point3Score")]
         public byte Point3Score
         {
-            [Obsolete("", true)] get => throw new NotSupportedException();
+            [Obsolete("", true), EditorBrowsable(EditorBrowsableState.Never)] get => throw new NotSupportedException();
             init => this._scores.Add(ScoreFormat.POINT_3, value);
         }
 
@@ -75,7 +74,7 @@ namespace PaperMalKing.AniList.Wrapper.Models
         public ulong Id { get; init; }
 
         [JsonPropertyName("customLists")]
-        public CustomList[]? CustomLists { get; init; } = Array.Empty<CustomList>();
+        public IReadOnlyList<CustomList>? CustomLists { get; init; } = Array.Empty<CustomList>();
 
         public string GetScore(ScoreFormat scoreFormat)
         {
