@@ -1,4 +1,5 @@
 ﻿#region LICENSE
+
 // PaperMalKing.
 // Copyright (C) 2021 N0D4N
 // 
@@ -14,6 +15,7 @@
 // 
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 #endregion
 
 using System.Text.Json.Serialization;
@@ -22,21 +24,21 @@ using PaperMalKing.AniList.Wrapper.Models.Interfaces;
 
 namespace PaperMalKing.AniList.Wrapper.Models
 {
-    public sealed class Review : ISiteUrlable
-    {
-        [JsonPropertyName("createdAt")]
-        public long CreatedAtTimeStamp { get; init; }
+	public sealed class Review : ISiteUrlable
+	{
+		[JsonPropertyName("createdAt")]
+		public long CreatedAtTimeStamp { get; init; }
 
-        [JsonPropertyName("siteUrl")]
-        public string Url { get; init; } = null!;
+		[JsonPropertyName("siteUrl")]
+		public string Url { get; init; } = null!;
 
-        [JsonPropertyName("summary")]
-        public string? Summary { get; init; }
+		[JsonPropertyName("summary")]
+		public string? Summary { get; init; }
 
-        [JsonPropertyName("media")]
-        public Media Media { get; init; } = null!;
-        
-        [JsonPropertyName("format")]
-        public MediaFormat Format { get; init; }
-    }
+		[JsonPropertyName("media")]
+		public Media Media { get; init; } = null!;
+
+		[JsonPropertyName("format")]
+		public MediaFormat Format { get; init; }
+	}
 }
