@@ -1,4 +1,5 @@
 ﻿#region LICENSE
+
 // PaperMalKing.
 // Copyright (C) 2021 N0D4N
 // 
@@ -14,27 +15,28 @@
 // 
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 #endregion
 
 using System;
 
 namespace PaperMalKing.Common.Attributes
 {
-    [AttributeUsage(AttributeTargets.Field, AllowMultiple = false, Inherited = false)]
-    public sealed class FeatureDescriptionAttribute : Attribute
-    {
-        public FeatureDescriptionAttribute(string description, string summary)
-        {
-            if(string.IsNullOrWhiteSpace(description))
-                throw new ArgumentException(null, nameof(description));
-            if(string.IsNullOrWhiteSpace(summary))
-                throw new ArgumentException(null, nameof(summary)); ;
-            this.Description = description;
-            this.Summary = summary;
-        }
+	[AttributeUsage(AttributeTargets.Field, AllowMultiple = false, Inherited = false)]
+	public sealed class FeatureDescriptionAttribute : Attribute
+	{
+		public FeatureDescriptionAttribute(string description, string summary)
+		{
+			if (string.IsNullOrWhiteSpace(description))
+				throw new ArgumentException(null, nameof(description));
+			if (string.IsNullOrWhiteSpace(summary))
+				throw new ArgumentException(null, nameof(summary));
+			this.Description = description;
+			this.Summary = summary;
+		}
 
-        public string Description { get; }
-        
-        public string Summary { get; }
-    }
+		public string Description { get; }
+
+		public string Summary { get; }
+	}
 }
