@@ -17,8 +17,8 @@
 #endregion
 
 using System;
+using System.Threading.RateLimiting;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using PaperMalKing.Common.Options;
 
@@ -26,18 +26,17 @@ namespace PaperMalKing.Common.RateLimiters
 {
 	public static class RateLimiterExtensions
 	{
-		public static RateLimiterHttpMessageHandler<T> ToHttpMessageHandler<T>(this IRateLimiter<T> rateLimiter) =>
+		public static RateLimiterHttpMessageHandler ToHttpMessageHandler(this RateLimiter rateLimiter) =>
 			new(rateLimiter);
 
-		public static IRateLimiter<T> ToRateLimiter<T>(this RateLimit rateLimit) =>
+		public static RateLimiter<T> ToRateLimiter<T>(this RateLimit rateLimit) =>
 			RateLimiterFactory.Create<T>(rateLimit);
 
-		public static IRateLimiter<T> ConfigurationLambda<TO, T>(IServiceProvider servicesProvider)
+		public static RateLimiter<T> ConfigurationLambda<TO, T>(IServiceProvider servicesProvider)
 		where TO : class, IRateLimitOptions<T>
 		{
-			var logger = servicesProvider.GetRequiredService<ILogger<IRateLimiter<T>>>();
 			var options = servicesProvider.GetRequiredService<IOptions<TO>>();
-			return options.Value.ToRateLimiter(logger);
+			return options.Value.ToRateLimiter();
 		}
 	}
 }
