@@ -1,4 +1,5 @@
 ﻿#region LICENSE
+
 // PaperMalKing.
 // Copyright (C) 2021 N0D4N
 // 
@@ -14,53 +15,47 @@
 // 
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 #endregion
 
-using System;
+using System.Diagnostics;
 using System.Net.Http;
 using System.Threading;
+using System.Threading.RateLimiting;
 using System.Threading.Tasks;
 
 namespace PaperMalKing.Common.RateLimiters
 {
-	public sealed class RateLimiterHttpMessageHandler<T> : DelegatingHandler
+	public sealed class RateLimiterHttpMessageHandler : DelegatingHandler
 	{
-		private IRateLimiter<T> _rateLimiter;
-		private readonly object _lock = new();
+		public RateLimiter RateLimiter { get; }
 
-		public IRateLimiter<T> RateLimiter
+		internal RateLimiterHttpMessageHandler(RateLimiter rateLimiter)
 		{
-			get => this._rateLimiter;
-			set
+			this.RateLimiter = rateLimiter;
+		}
+
+		protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+		{
+			while (!cancellationToken.IsCancellationRequested)
 			{
-				var old = this.RateLimiter;
-				lock (this._lock)
+				using (var rateLimitLease = await this.RateLimiter.AcquireAsync(1, cancellationToken).ConfigureAwait(false))
 				{
-					this._rateLimiter = value;
+					if (rateLimitLease.IsAcquired)
+					{
+						return await base.SendAsync(request, cancellationToken).ConfigureAwait(false);
+					}
 				}
-
-				if (old is IDisposable disposable)
-					disposable.Dispose();
 			}
+
+			cancellationToken.ThrowIfCancellationRequested();
+			throw new UnreachableException();
 		}
 
-		internal RateLimiterHttpMessageHandler(IRateLimiter<T> rateLimiter)
+		protected override void Dispose(bool disposing)
 		{
-			this._rateLimiter = rateLimiter;
-		}
-
-		protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request,
-																	 CancellationToken cancellationToken)
-		{
-			await this.RateLimiter.TickAsync(cancellationToken).ConfigureAwait(false);
-			return await base.SendAsync(request, cancellationToken).ConfigureAwait(false);
-		}
-
-		public new void Dispose()
-		{
-			base.Dispose();
-			if (this.RateLimiter is IDisposable disposable)
-				disposable.Dispose();
+			RateLimiter.Dispose();
+			base.Dispose(disposing);
 		}
 	}
 }
