@@ -22,8 +22,10 @@ namespace PaperMalKing.Database.Migrations
 {
     public partial class InitialMigration : Migration
     {
+		#pragma warning disable MA0051
         protected override void Up(MigrationBuilder migrationBuilder)
-        {
+		#pragma warning restore MA0051
+		{
             migrationBuilder.CreateTable(
                 name: "BotUsers",
                 columns: table => new
