@@ -36,41 +36,37 @@ namespace PaperMalKing.UpdatesProviders.MyAnimeList
 {
 	internal sealed class MalUpdateProviderConfigurator : IUpdateProviderConfigurator<MalUpdateProvider>
 	{
-		/// <inheritdoc />
-		public void ConfigureNonStatic(IConfiguration configuration, IServiceCollection serviceCollection)
-		{ }
-
-		public static void Configure(IConfiguration configuration, IServiceCollection services)
+		public static void Configure(IConfiguration configuration, IServiceCollection serviceCollection)
 		{
-			services.AddOptions<MalOptions>().Bind(configuration.GetSection(Constants.Name));
-			services.AddSingleton(provider => RateLimiterExtensions.ConfigurationLambda<MalOptions, MyAnimeListClient>(provider));
+			serviceCollection.AddOptions<MalOptions>().Bind(configuration.GetSection(Constants.Name));
+			serviceCollection.AddSingleton<RateLimiter<MyAnimeListClient>>(RateLimiterExtensions.ConfigurationLambda<MalOptions, MyAnimeListClient>);
 
 			var retryPolicy = HttpPolicyExtensions.HandleTransientHttpError()
 												  .OrResult(message => message.StatusCode == HttpStatusCode.TooManyRequests)
 												  .WaitAndRetryAsync(Backoff.DecorrelatedJitterBackoffV2(TimeSpan.FromSeconds(10), 5));
-			services.AddHttpClient(MalOptions.MyAnimeList).AddPolicyHandler(retryPolicy).ConfigurePrimaryHttpMessageHandler(_ => new HttpClientHandler
+			serviceCollection.AddHttpClient(MalOptions.MyAnimeList).AddPolicyHandler(retryPolicy).ConfigurePrimaryHttpMessageHandler(_ => new HttpClientHandler
 			{
 				UseCookies = true,
 				CookieContainer = new()
 			}).AddHttpMessageHandler(provider =>
 			{
-				var rl = provider.GetRequiredService<IRateLimiter<MyAnimeListClient>>();
+				var rl = provider.GetRequiredService<RateLimiter<MyAnimeListClient>>();
 				return rl.ToHttpMessageHandler();
 			}).ConfigureHttpClient(client =>
 			{
 				client.DefaultRequestHeaders.UserAgent.Clear();
 				client.DefaultRequestHeaders.UserAgent.ParseAdd("Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:76.0) Gecko/20100101 Firefox/76.0");
 			});
-			services.AddSingleton<MyAnimeListClient>(provider =>
+			serviceCollection.AddSingleton<MyAnimeListClient>(provider =>
 			{
 				var factory = provider.GetRequiredService<IHttpClientFactory>();
 				var logger = provider.GetRequiredService<ILogger<MyAnimeListClient>>();
 				return new(logger, factory.CreateClient(MalOptions.MyAnimeList));
 			});
-			services.AddSingleton<IExecuteOnStartupService, MalExecuteOnStartupService>();
-			services.AddSingleton<IUserFeaturesService<MalUserFeatures>,MalUserFeaturesService>();
-			services.AddSingleton<MalUserService>();
-			services.AddSingleton<IUpdateProvider, MalUpdateProvider>();
+			serviceCollection.AddSingleton<IExecuteOnStartupService, MalExecuteOnStartupService>();
+			serviceCollection.AddSingleton<IUserFeaturesService<MalUserFeatures>,MalUserFeaturesService>();
+			serviceCollection.AddSingleton<MalUserService>();
+			serviceCollection.AddSingleton<IUpdateProvider, MalUpdateProvider>();
 		}
 	}
 }
