@@ -22,7 +22,7 @@ namespace PaperMalKing.MyAnimeList.Wrapper.Models.List.Types
 {
 	internal interface IListType<T> where T : class, IListEntry
 	{
-		ListEntryType ListEntryType { get; }
-		string LatestUpdatesUrl(string username);
+		abstract static ListEntryType ListEntryType { get; }
+		abstract static string LatestUpdatesUrl(string username);
 	}
 }
