@@ -20,10 +20,10 @@ using PaperMalKing.Common.Enums;
 
 namespace PaperMalKing.MyAnimeList.Wrapper.Models.Rss.Types
 {
-	internal readonly struct AnimeRssFeed : IRssFeedType
+	internal abstract class AnimeRssFeed : IRssFeedType
 	{
-		public string Url => Constants.RSS_ANIME_URL;
+		public static string Url => Constants.RSS_ANIME_URL;
 
-		public ListEntryType Type => ListEntryType.Anime;
+		public static ListEntryType Type => ListEntryType.Anime;
 	}
 }
