@@ -28,14 +28,13 @@ using PaperMalKing.UpdatesProviders.Base.Exceptions;
 
 namespace PaperMalKing.UpdatesProviders.Base
 {
-	[SuppressMessage("Microsoft.Design", "CA1051")]
-	public abstract class BaseUpdateProviderUserCommandsModule : BaseCommandModule
+	[SuppressMessage("Style", "VSTHRD200:Use \"Async\" suffix for async methods")]
+	public abstract class BaseUpdateProviderUserCommandsModule<TUpdateProviderUserService> : BaseCommandModule where TUpdateProviderUserService: class, IUpdateProviderUserService
 	{
-		protected readonly ILogger<BaseUpdateProviderUserCommandsModule> Logger;
-		protected readonly IUpdateProviderUserService UserService;
+		protected ILogger<BaseUpdateProviderUserCommandsModule<TUpdateProviderUserService>> Logger { get; }
+		protected TUpdateProviderUserService UserService { get; }
 
-		/// <inheritdoc />
-		protected BaseUpdateProviderUserCommandsModule(IUpdateProviderUserService userService, ILogger<BaseUpdateProviderUserCommandsModule> logger)
+		protected BaseUpdateProviderUserCommandsModule(TUpdateProviderUserService userService, ILogger<BaseUpdateProviderUserCommandsModule<TUpdateProviderUserService>> logger)
 		{
 			this.UserService = userService;
 			this.Logger = logger;
@@ -44,7 +43,7 @@ namespace PaperMalKing.UpdatesProviders.Base
 		public virtual async Task AddUserCommand(CommandContext ctx, [RemainingText] [Description("Your username")]
 												 string username)
 		{
-			this.Logger.LogInformation("Trying to add {ProviderUsername} {Member} to {Name} update provider",username, ctx.Member,  UserService.Name);
+			this.Logger.LogInformation("Trying to add {ProviderUsername} {Member} to {Name} update provider",username, ctx.Member,  TUpdateProviderUserService.Name);
 			BaseUser user;
 			try
 			{
@@ -54,20 +53,20 @@ namespace PaperMalKing.UpdatesProviders.Base
 			{
 				var embed = ex is UserProcessingException ? EmbedTemplate.ErrorEmbed(ctx, ex.Message) : EmbedTemplate.UnknownErrorEmbed(ctx);
 				await ctx.RespondAsync(embed: embed.Build()).ConfigureAwait(false);
-				this.Logger.LogError(ex,"Failed to add {ProviderUsername} {Member} to {Name} update provider",username, ctx.Member,  UserService.Name);
+				this.Logger.LogError(ex,"Failed to add {ProviderUsername} {Member} to {Name} update provider",username, ctx.Member,  TUpdateProviderUserService.Name);
 				throw;
 			}
 
-			this.Logger.LogInformation("Successfully added {ProviderUsername} {Member} to {Name} update provider",username, ctx.Member,  UserService.Name);
+			this.Logger.LogInformation("Successfully added {ProviderUsername} {Member} to {Name} update provider",username, ctx.Member,  TUpdateProviderUserService.Name);
 
 			await ctx.RespondAsync(embed: EmbedTemplate.SuccessEmbed(ctx,
-				$"Successfully added {user.Username} to {this.UserService.Name} update checker")).ConfigureAwait(false);
+				$"Successfully added {user.Username} to {TUpdateProviderUserService.Name} update checker")).ConfigureAwait(false);
 		}
 
 
 		public virtual async Task RemoveUserInGuildCommand(CommandContext ctx)
 		{
-			this.Logger.LogInformation("Trying to remove {Member} from {Name} update provider", ctx.Member, UserService.Name);
+			this.Logger.LogInformation("Trying to remove {Member} from {Name} update provider", ctx.Member, TUpdateProviderUserService.Name);
 			BaseUser user;
 			try
 			{
@@ -77,14 +76,14 @@ namespace PaperMalKing.UpdatesProviders.Base
 			{
 				var embed = ex is UserProcessingException ? EmbedTemplate.ErrorEmbed(ctx, ex.Message) : EmbedTemplate.UnknownErrorEmbed(ctx);
 				await ctx.RespondAsync(embed: embed).ConfigureAwait(false);
-				this.Logger.LogError(ex,"Failed to remove {Member} from {Name} update provider",ctx.Member,  UserService.Name);
+				this.Logger.LogError(ex,"Failed to remove {Member} from {Name} update provider",ctx.Member,  TUpdateProviderUserService.Name);
 
 				throw;
 			}
-			this.Logger.LogInformation("Successfully removed {Member} from {Name} update provider", ctx.Member, UserService.Name);
+			this.Logger.LogInformation("Successfully removed {Member} from {Name} update provider", ctx.Member, TUpdateProviderUserService.Name);
 
 			await ctx.RespondAsync(embed: EmbedTemplate.SuccessEmbed(ctx,
-				$"Successfully removed {user.Username} from {this.UserService.Name} update checker")).ConfigureAwait(false);
+				$"Successfully removed {user.Username} from {TUpdateProviderUserService.Name} update checker")).ConfigureAwait(false);
 		}
 
 		public virtual async Task RemoveUserHereCommand(CommandContext ctx)
@@ -109,7 +108,7 @@ namespace PaperMalKing.UpdatesProviders.Base
 			try
 			{
 				var i = 1;
-				await foreach (var user in this.UserService.ListUsersAsync(ctx.Guild.Id))
+				await foreach (var user in this.UserService.ListUsersAsync(ctx.Guild.Id).ConfigureAwait(false))
 				{
 					if (sb.Length + user.Username.Length > 2048)
 					{
