@@ -38,14 +38,14 @@ namespace PaperMalKing.Commands
 {
 	[ModuleLifespan(ModuleLifespan.Singleton)]
 	[SuppressMessage("Performance", "CA1822")]
+	[SuppressMessage("Style", "VSTHRD200:Use \"Async\" suffix for async methods")]
 	public sealed class UngroupedCommands : BaseCommandModule
 	{
 		private readonly UpdateProvidersConfigurationService _providersConfigurationService;
 		private readonly IHostApplicationLifetime _lifetime;
 
-		private DiscordEmbed? AboutEmbed;
+		private static DiscordEmbed? AboutEmbed;
 
-		/// <inheritdoc />
 		public UngroupedCommands(UpdateProvidersConfigurationService providersConfigurationService, IHostApplicationLifetime lifetime)
 		{
 			this._providersConfigurationService = providersConfigurationService;
@@ -54,8 +54,7 @@ namespace PaperMalKing.Commands
 
 		[Command("say")]
 		[Description("Sends embed in selected channel with selected text")]
-		[OwnerOrPermission(Permissions.ManageGuild)]
-		[SuppressMessage("Microsoft.Design", "CA1031")]
+		[OwnerOrPermissions(Permissions.ManageGuild)]
 		public async Task SayCommand(CommandContext context, [Description("Channel where the embed will be send")]
 									 DiscordChannel channelToSayIn, [RemainingText, Description("Text to send")]
 									 string messageContent)
@@ -73,7 +72,9 @@ namespace PaperMalKing.Commands
 			{
 				await channelToSayIn.SendMessageAsync(embed: embed).ConfigureAwait(false);
 			}
+			#pragma warning disable CA1031
 			catch
+			#pragma warning restore CA1031
 			{
 				await context.RespondAsync("Couldn't send message. Check permissions for bot and try again.").ConfigureAwait(false);
 			}
@@ -82,9 +83,9 @@ namespace PaperMalKing.Commands
 		[Command("About")]
 		[Description("Displays info about bot")]
 		[Aliases("Info")]
-		public async Task AboutCommand(CommandContext context)
+		public Task AboutCommand(CommandContext context)
 		{
-			if (this.AboutEmbed == null)
+			if (AboutEmbed == null)
 			{
 				var botVersion = Assembly.GetEntryAssembly()?.GetName().Version?.ToString(3) ?? "";
 				var dotnetVersion = Environment.Version.ToString(3);
@@ -107,10 +108,10 @@ namespace PaperMalKing.Commands
 				}.WithThumbnail(context.Client.CurrentUser.AvatarUrl);
 				embedBuilder.AddField("Links", links, true);
 				embedBuilder.AddField("Versions", versions, true);
-				Interlocked.Exchange(ref this.AboutEmbed, embedBuilder.Build());
+				Interlocked.Exchange(ref AboutEmbed, embedBuilder.Build());
 			}
 
-			await context.RespondAsync(embed: this.AboutEmbed).ConfigureAwait(false);
+			return context.RespondAsync(embed: AboutEmbed);
 		}
 
 		[Command("DeleteMessages")]
