@@ -18,13 +18,15 @@
 
 #endregion
 
+using System.Collections.Generic;
+
 namespace PaperMalKing.Options
 {
 	public sealed class CommandsOptions
 	{
 		public const string Commands = "Discord:Commands";
 
-		public string[] Prefixes { get; set; } = null!;
+		public IReadOnlyList<string> Prefixes { get; set; } = null!;
 
 		public bool EnableMentionPrefix { get; set; }
 
