@@ -20,7 +20,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
@@ -38,21 +37,18 @@ namespace PaperMalKing.Services
 	public sealed class CommandsService : ICommandsService
 	{
 		private readonly ILogger<CommandsService> _logger;
-		private readonly CommandsOptions _options;
 		public CommandsNextExtension CommandsExtension { get; }
 
-		[SuppressMessage("Microsoft.Design", "CA1031")]
 		public CommandsService(IOptions<CommandsOptions> options, IServiceProvider provider, DiscordClient client, ILogger<CommandsService> logger)
 		{
 			this._logger = logger;
 			this._logger.LogTrace("Building Commands service");
-			this._options = options.Value;
 
 			var config = new CommandsNextConfiguration
 			{
-				CaseSensitive = this._options.CaseSensitive,
-				EnableMentionPrefix = this._options.EnableMentionPrefix,
-				StringPrefixes = this._options.Prefixes,
+				CaseSensitive = options.Value.CaseSensitive,
+				EnableMentionPrefix = options.Value.EnableMentionPrefix,
+				StringPrefixes = options.Value.Prefixes,
 				DmHelp = false,
 				Services = provider
 			};
@@ -70,8 +66,8 @@ namespace PaperMalKing.Services
 			this.CommandsExtension.RegisterUserFriendlyTypeName<ushort>("unsigned small integer");
 			this.CommandsExtension.RegisterUserFriendlyTypeName<uint>("unsigned integer");
 			this.CommandsExtension.RegisterUserFriendlyTypeName<ulong>("unsigned integer");
-			this.CommandsExtension.CommandErrored += this.CommandsExtensionOnCommandErrored;
-			this.CommandsExtension.CommandExecuted += this.CommandsExtensionOnCommandExecuted;
+			this.CommandsExtension.CommandErrored += this.CommandsExtensionOnCommandErroredAsync;
+			this.CommandsExtension.CommandExecuted += this.CommandsExtensionOnCommandExecutedAsync;
 
 			var assemblies = Utils.LoadAndListPmkAssemblies();
 			HashSet<Type> nestedTypesNotToRegister = new();
@@ -95,7 +91,9 @@ namespace PaperMalKing.Services
 
 						this.CommandsExtension.RegisterCommands(type);
 					}
+					#pragma warning disable CA1031
 					catch (Exception ex)
+					#pragma warning restore CA1031
 					{
 						this._logger.LogError(ex, "Error occured while trying to register {FullName}", type.FullName);
 					}
@@ -110,13 +108,13 @@ namespace PaperMalKing.Services
 			this._logger.LogTrace("Building Commands service");
 		}
 
-		private Task CommandsExtensionOnCommandExecuted(CommandsNextExtension sender, CommandExecutionEventArgs e)
+		private Task CommandsExtensionOnCommandExecutedAsync(CommandsNextExtension sender, CommandExecutionEventArgs e)
 		{
 			this._logger.LogDebug("{Command} was successfully executed by request of {Member}", e.Command, e.Context.Member);
 			return Task.CompletedTask;
 		}
 
-		private Task CommandsExtensionOnCommandErrored(CommandsNextExtension sender, CommandErrorEventArgs e)
+		private Task CommandsExtensionOnCommandErroredAsync(CommandsNextExtension sender, CommandErrorEventArgs e)
 		{
 			if (e.Exception is CommandNotFoundException ex)
 			{
